requirements.txt
pyserial==3.5
python_weather==1.0.3
